pyproject.toml
[project]
name = "comfyui-secnodes"
description = "Self-contained nodes for SeC (Segment Concept) - SOTA video object segmentation outperforming SAM 2.1"
version = "1.2.2"
license = { file = "LICENSE" }
dependencies = [
    "huggingface_hub>=0.20.0",
    "peft>=0.10.0",
    "timm>=0.9.0",
    "pillow>=8.0.0",
    "opencv-python-headless>=4.5.0",
    "scikit-image>=0.19.0",
    "scipy>=1.7.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.1.0",
    "ninja>=1.11.0",
    "iopath",
    "packaging>=21.0"
]

[tool.comfy]
PublisherId = "9nate-drake"
DisplayName = "ComfyUI-SecNodes"
Icon = ""

[project.urls]
Repository = "https://github.com/9nate-drake/Comfyui-SecNodes"
